+++ airbyte-integrations/bases/source-acceptance-test/unit_tests/test_backward_compatibility.py
@@ -3,30 +3,33 @@
 #
 
 from dataclasses import dataclass
+from typing import MutableMapping, Union
 
 import pytest
-from airbyte_cdk.models import ConnectorSpecification
+from airbyte_cdk.models import AirbyteStream, ConnectorSpecification
+from source_acceptance_test.tests.test_core import TestDiscovery as _TestDiscovery
 from source_acceptance_test.tests.test_core import TestSpec as _TestSpec
-from source_acceptance_test.utils.backward_compatibility import NonBackwardCompatibleSpecError
+from source_acceptance_test.utils.backward_compatibility import NonBackwardCompatibleError, validate_previous_configs
 
 from .conftest import does_not_raise
 
 
 @dataclass
-class SpecTransition:
+class Transition:
     """An helper class to improve readability of the test cases"""
 
-    previous_connector_specification: ConnectorSpecification
-    current_connector_specification: ConnectorSpecification
+    previous: Union[ConnectorSpecification, MutableMapping[str, AirbyteStream]]
+    current: Union[ConnectorSpecification, MutableMapping[str, AirbyteStream]]
     should_fail: bool
     name: str
+    is_valid_json_schema: bool = True
 
     def as_pytest_param(self):
-        return pytest.param(self.previous_connector_specification, self.current_connector_specification, self.should_fail, id=self.name)
+        return pytest.param(self.previous, self.current, self.should_fail, id=self.name)
 
 
 FAILING_SPEC_TRANSITIONS = [
-    SpecTransition(
+    Transition(
         ConnectorSpecification(connectionSpecification={}),
         ConnectorSpecification(
             connectionSpecification={
@@ -36,7 +39,7 @@ FAILING_SPEC_TRANSITIONS = [
         should_fail=True,
         name="Top level: declaring the required field should fail.",
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -58,7 +61,7 @@ FAILING_SPEC_TRANSITIONS = [
         should_fail=True,
         name="Nested level: adding the required field should fail.",
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "required": ["a"],
@@ -72,7 +75,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Top level: adding a new required property should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -103,7 +106,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Nested level: adding a new required property should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -123,7 +126,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Nullable: Making a field not nullable should fail (not in a list).",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -143,7 +146,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Nested level: Narrowing a field type should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -163,7 +166,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Nullable field: Making a field not nullable should fail",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -183,7 +186,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Changing a field type should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -203,7 +206,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Changing a field type from a string to a list with a different type value should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -223,7 +226,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Changing a field type should fail from a list to string with different value should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -243,7 +246,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Changing a field type in list should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -263,7 +266,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Making a field nullable and changing type should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -283,7 +286,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Making a field nullable and changing type should fail (change list order).",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -303,7 +306,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Nullable field: Changing a field type should fail",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -333,7 +336,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Changing a field type in oneOf should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -363,7 +366,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Narrowing a field type in oneOf should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -383,7 +386,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Top level: Narrowing a field enum should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -403,7 +406,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Nested level: Narrowing a field enum should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -423,7 +426,7 @@ FAILING_SPEC_TRANSITIONS = [
         name="Top level: Declaring a field enum should fail.",
         should_fail=True,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -443,10 +446,52 @@ FAILING_SPEC_TRANSITIONS = [
         name="Nested level: Declaring a field enum should fail.",
         should_fail=True,
     ),
+    Transition(
+        ConnectorSpecification(
+            connectionSpecification={
+                "type": "object",
+                "properties": {
+                    "my_string": {"type": "string"},
+                },
+            }
+        ),
+        ConnectorSpecification(
+            connectionSpecification={
+                "type": "object",
+                "properties": {
+                    "my_string": {"type": {}},
+                },
+            }
+        ),
+        name="Changing a 'type' field from a string to something else than a list should fail.",
+        should_fail=True,
+        is_valid_json_schema=False,
+    ),
+    Transition(
+        ConnectorSpecification(
+            connectionSpecification={
+                "type": "object",
+                "properties": {
+                    "my_string": {"type": ["string"]},
+                },
+            }
+        ),
+        ConnectorSpecification(
+            connectionSpecification={
+                "type": "object",
+                "properties": {
+                    "my_string": {"type": {}},
+                },
+            }
+        ),
+        name="Changing a 'type' field from a list to something else than a string should fail.",
+        should_fail=True,
+        is_valid_json_schema=False,
+    ),
 ]
 
 VALID_SPEC_TRANSITIONS = [
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -468,7 +513,55 @@ VALID_SPEC_TRANSITIONS = [
         name="Not changing a spec should not fail",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
+        ConnectorSpecification(
+            connectionSpecification={
+                "type": "object",
+                "required": ["my_required_string"],
+                "additionalProperties": False,
+                "properties": {
+                    "my_required_string": {"type": "string"},
+                },
+            }
+        ),
+        ConnectorSpecification(
+            connectionSpecification={
+                "type": "object",
+                "required": ["my_required_string"],
+                "additionalProperties": True,
+                "properties": {
+                    "my_required_string": {"type": "string"},
+                },
+            }
+        ),
+        name="Top level: Changing the value of additionalProperties should not fail",
+        should_fail=False,
+    ),
+    Transition(
+        ConnectorSpecification(
+            connectionSpecification={
+                "type": "object",
+                "properties": {
+                    "my_nested_object": {"type": "object", "properties": {"my_property": {"type": ["integer"]}}},
+                },
+            }
+        ),
+        ConnectorSpecification(
+            connectionSpecification={
+                "type": "object",
+                "properties": {
+                    "my_nested_object": {
+                        "type": "object",
+                        "additionalProperties": True,
+                        "properties": {"my_property": {"type": ["integer"]}},
+                    },
+                },
+            }
+        ),
+        name="Nested level: Changing the value of additionalProperties should not fail",
+        should_fail=False,
+    ),
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -491,7 +584,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Adding an optional field should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -518,7 +611,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Adding an optional object with required properties should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -538,7 +631,7 @@ VALID_SPEC_TRANSITIONS = [
         name="No change should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -558,7 +651,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Changing a field type from a list to a string with same value should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -578,7 +671,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Changing a field type from a string to a list should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -598,7 +691,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Adding a field type in list should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -618,7 +711,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Making a field nullable should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -638,7 +731,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Making a field nullable should not fail (change list order).",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -658,7 +751,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Making a field nullable should not fail (from a list).",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -678,7 +771,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Making a field nullable should not fail (from a list, changing order).",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -698,7 +791,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Nullable field: Changing order should not fail",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -718,7 +811,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Nested level: Expanding a field type should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -748,7 +841,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Changing a order in oneOf should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -768,7 +861,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Top level: Expanding a field enum should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -788,7 +881,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Nested level: Expanding a field enum should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -806,7 +899,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Top level: Adding a new optional field with enum should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -826,7 +919,7 @@ VALID_SPEC_TRANSITIONS = [
         name="Top level: Removing the field enum should not fail.",
         should_fail=False,
     ),
-    SpecTransition(
+    Transition(
         ConnectorSpecification(
             connectionSpecification={
                 "type": "object",
@@ -851,15 +944,321 @@ VALID_SPEC_TRANSITIONS = [
 # Checking that all transitions in FAILING_SPEC_TRANSITIONS have should_fail == True to prevent typos
 assert all([transition.should_fail for transition in FAILING_SPEC_TRANSITIONS])
 # Checking that all transitions in VALID_SPEC_TRANSITIONS have should_fail = False to prevent typos
-assert not all([transition.should_fail for transition in VALID_SPEC_TRANSITIONS])
-
+assert all([not transition.should_fail for transition in VALID_SPEC_TRANSITIONS])
 
 ALL_SPEC_TRANSITIONS_PARAMS = [transition.as_pytest_param() for transition in FAILING_SPEC_TRANSITIONS + VALID_SPEC_TRANSITIONS]
 
 
 @pytest.mark.parametrize("previous_connector_spec, actual_connector_spec, should_fail", ALL_SPEC_TRANSITIONS_PARAMS)
-def test_backward_compatibility(previous_connector_spec, actual_connector_spec, should_fail):
+def test_spec_backward_compatibility(previous_connector_spec, actual_connector_spec, should_fail):
     t = _TestSpec()
-    expectation = pytest.raises(NonBackwardCompatibleSpecError) if should_fail else does_not_raise()
+    expectation = pytest.raises(NonBackwardCompatibleError) if should_fail else does_not_raise()
     with expectation:
         t.test_backward_compatibility(False, actual_connector_spec, previous_connector_spec, 10)
+
+
+VALID_JSON_SCHEMA_TRANSITIONS_PARAMS = [
+    transition.as_pytest_param() for transition in FAILING_SPEC_TRANSITIONS + VALID_SPEC_TRANSITIONS if transition.is_valid_json_schema
+]
+
+
+@pytest.mark.slow
+@pytest.mark.parametrize("previous_connector_spec, actual_connector_spec, should_fail", VALID_JSON_SCHEMA_TRANSITIONS_PARAMS)
+def test_validate_previous_configs(previous_connector_spec, actual_connector_spec, should_fail):
+    expectation = pytest.raises(NonBackwardCompatibleError) if should_fail else does_not_raise()
+    with expectation:
+        validate_previous_configs(previous_connector_spec, actual_connector_spec, 200)
+
+
+FAILING_CATALOG_TRANSITIONS = [
+    Transition(
+        name="Removing a stream from a catalog should fail.",
+        should_fail=True,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            ),
+            "other_test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "other_test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            ),
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            )
+        },
+    ),
+    Transition(
+        name="Changing a field type should fail.",
+        should_fail=True,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            )
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "integer"}}}}},
+                }
+            )
+        },
+    ),
+    Transition(
+        name="Renaming a stream should fail.",
+        should_fail=True,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            )
+        },
+        current={
+            "new_test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "new_test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            )
+        },
+    ),
+    Transition(
+        name="Changing a cursor in a stream should fail.",
+        should_fail=True,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["a"],
+                }
+            ),
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["b"],
+                }
+            ),
+        },
+    ),
+    Transition(
+        name="Changing a cursor in a stream should fail (nested cursors).",
+        should_fail=True,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["a"],
+                }
+            ),
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["a", "b"],
+                }
+            ),
+        },
+    ),
+    Transition(
+        name="Changing a cursor in a stream should fail (nested cursors removal).",
+        should_fail=True,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["a", "b"],
+                }
+            ),
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["a"],
+                }
+            ),
+        },
+    ),
+    Transition(
+        name="Adding a stream but changing cursor should fail.",
+        should_fail=True,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["a"],
+                }
+            ),
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["b"],
+                }
+            ),
+            "other_test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "other_test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            ),
+        },
+    ),
+]
+
+VALID_CATALOG_TRANSITIONS = [
+    Transition(
+        name="Adding a stream to a catalog should not fail.",
+        should_fail=False,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            )
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            ),
+            "other_test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "other_test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            ),
+        },
+    ),
+    Transition(
+        name="Making a field nullable should not fail.",
+        should_fail=False,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            )
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": ["string", "null"]}}}}},
+                }
+            )
+        },
+    ),
+    Transition(
+        name="Changing 'type' field to list should not fail.",
+        should_fail=False,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            )
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": ["string"]}}}}},
+                }
+            )
+        },
+    ),
+    Transition(
+        name="Removing a field should not fail.",
+        should_fail=False,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {
+                        "properties": {
+                            "user": {"type": "object", "properties": {"username": {"type": "string"}, "email": {"type": "string"}}}
+                        }
+                    },
+                }
+            )
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                }
+            )
+        },
+    ),
+    Transition(
+        name="Not changing a cursor in a stream should not fail.",
+        should_fail=False,
+        previous={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["a"],
+                }
+            ),
+        },
+        current={
+            "test_stream": AirbyteStream.parse_obj(
+                {
+                    "name": "test_stream",
+                    "json_schema": {"properties": {"user": {"type": "object", "properties": {"username": {"type": "string"}}}}},
+                    "default_cursor_field": ["a"],
+                }
+            ),
+        },
+    ),
+]
+
+# Checking that all transitions in FAILING_CATALOG_TRANSITIONS have should_fail == True to prevent typos
+assert all([transition.should_fail for transition in FAILING_CATALOG_TRANSITIONS])
+# Checking that all transitions in VALID_CATALOG_TRANSITIONS have should_fail = False to prevent typos
+assert all([not transition.should_fail for transition in VALID_CATALOG_TRANSITIONS])
+
+ALL_CATALOG_TRANSITIONS_PARAMS = [transition.as_pytest_param() for transition in FAILING_CATALOG_TRANSITIONS + VALID_CATALOG_TRANSITIONS]
+
+
+@pytest.mark.parametrize("previous_discovered_catalog, discovered_catalog, should_fail", ALL_CATALOG_TRANSITIONS_PARAMS)
+def test_catalog_backward_compatibility(previous_discovered_catalog, discovered_catalog, should_fail):
+    t = _TestDiscovery()
+    expectation = pytest.raises(NonBackwardCompatibleError) if should_fail else does_not_raise()
+    with expectation:
+        t.test_backward_compatibility(False, discovered_catalog, previous_discovered_catalog)