+++ airbyte-integrations/connectors/source-salesforce/source_salesforce/api.py
@@ -137,6 +137,7 @@ UNSUPPORTED_BULK_API_SALESFORCE_OBJECTS = [
     "DeclinedEventRelation",
     "EventWhoRelation",
     "FieldSecurityClassification",
+    "KnowledgeArticle",
     "OrderStatus",
     "PartnerRole",
     "QuoteTemplateRichTextData",